\name{CoINcIDE-package}
\alias{CoINcIDE-package}
\alias{CoINcIDE}
\docType{package}
\title{
\packageTitle{CoINcIDE}
}
\description{
\packageDescription{CoINcIDE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoINcIDE}
\packageIndices{CoINcIDE}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{CoINcIDE}

Maintainer: \packageMaintainer{CoINcIDE}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
cat("this is an example")
}
