\name{Coincide-package}
\alias{Coincide-package}
\alias{Coincide}
\docType{package}
\title{
\packageTitle{Coincide}
}
\description{
\packageDescription{Coincide}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Coincide}
\packageIndices{Coincide}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{Coincide}

Maintainer: \packageMaintainer{Coincide}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
cat("this is an example")
}
